backend/app/processors/image_processor.py
"""
Image processor implementation using Groq vision API
"""
from typing import Dict, Any, Optional
from PIL import Image
from io import BytesIO
from app.processors.base import BaseProcessor
from app.services.agent.llm.groq_client import GroqClient
from app.core.config import settings


class ImageProcessor(BaseProcessor):
    """Image processor using Groq vision API"""
    
    def __init__(self):
        # Note: Groq doesn't have vision API yet, this is a placeholder
        # In production, you might use OpenAI vision or another service
        self.groq_client = GroqClient()
    
    async def process(self, data: bytes, metadata: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
        """Process image file"""
        # Get image info
        image_info = await self.analyze(data)
        
        # Extract text (OCR-like functionality would go here)
        # For now, return image info
        text = await self.extract_text(data)
        
        return {
            "type": "image",
            "text": text,
            "info": image_info,
            "metadata": metadata or {}
        }
    
    async def extract_text(self, data: bytes) -> str:
        """Extract text from image (OCR)"""
        # This would integrate with OCR service or vision API
        # For now, return empty string
        # In production, use Groq vision or OpenAI vision API
        return ""
    
    async def analyze(self, data: bytes) -> Dict[str, Any]:
        """Analyze image properties"""
        image = Image.open(BytesIO(data))
        
        return {
            "width": image.width,
            "height": image.height,
            "format": image.format,
            "mode": image.mode
        }

