backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# HTTP client for BlueBubbles
httpx==0.27.2

# LLM
groq==0.11.0

# Twilio
twilio==9.3.5

# Google APIs
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0

# Notion API
notion-client==2.2.1

# Document processing
PyPDF2==3.0.1
pdfplumber==0.11.4
Pillow==11.0.0

# Utilities
pydantic==2.9.2
pydantic-settings==2.5.2
python-dateutil==2.9.0.post0

